+++ src/knowledge-memory/contracts/knowledge-ingestion-curation-profile.js
@@ -1,0 +1,56 @@
+export class KnowledgeIngestionCurationProfile {
+  constructor({
+    profileName,
+    objectives = [], sourceClasses = [], contentClasses = [], acquisitionModels = [],
+    intakeOutcomes = [], quarantineReasons = [], pipelineStages = [], chunkingStrategies = [],
+    qualityGates = [], processingStates = [], failureTypes = [], ingestionApis = [],
+    ingestionEvents = [], qualityAttributes = [], architecturalRules = [], architectureBoundaries = [],
+    registeredSourcesRequired = true, sourceWorkloadAuthenticated = true, intakeScopeBound = true,
+    idempotentAcquisition = true, synchronizationReconciled = true, deletionGoverned = true,
+    intakeValidated = true, unsafeContentQuarantined = true, processingIsolated = true,
+    originalPreserved = true, actualTypeDetected = true, structurePreserved = true,
+    extractionDerived = true, processingMethodsVersioned = true, normalizationMeaningPreserved = true,
+    metadataProvenance = true, classificationPropagated = true, sensitiveDataGoverned = true,
+    chunkBoundariesPreserved = true, contentAwareChunking = true, aiEnrichmentDerived = true,
+    curatorAuthorized = true, separationOfDuties = true, qualityGatesRequired = true,
+    publicationHandoffOnly = true, processingProvenance = true, reprocessingControlled = true,
+    retryBounded = true, poisonContentIsolated = true, tenantPropertyIsolation = true,
+    temporaryDataDeleted = true, providerGoverned = true, vendorNeutral = true,
+    technologyIndependent = true,
+    ingestionActivatesKnowledge = false, uploadGrantsPublication = false,
+    duplicateMergesScopes = false, failedSyncMeansDeletion = false, extensionDefinesType = false,
+    extractionIsSourceTruth = false, normalizationChangesMeaning = false,
+    classificationBecomesLessRestrictive = false, chunksCrossBoundaries = false,
+    oneChunkStrategy = false, aiChangesAuthority = false, aiPublishesItself = false,
+    curatorSelfApprovesCritical = false, readinessMeansActivation = false,
+    reprocessingReplacesActive = false, retriesUnsafeContent = false,
+    externalProviderDefinesMeaning = false, directDatabaseAccess = false,
+    selectsProcessingProvider = false
+  }) {
+    this.profileName = profileName;
+    for (const [key, value] of Object.entries({
+      objectives, sourceClasses, contentClasses, acquisitionModels, intakeOutcomes,
+      quarantineReasons, pipelineStages, chunkingStrategies, qualityGates, processingStates,
+      failureTypes, ingestionApis, ingestionEvents, qualityAttributes, architecturalRules,
+      architectureBoundaries
+    })) this[key] = Object.freeze([...value]);
+    for (const [key, value] of Object.entries({
+      registeredSourcesRequired, sourceWorkloadAuthenticated, intakeScopeBound,
+      idempotentAcquisition, synchronizationReconciled, deletionGoverned, intakeValidated,
+      unsafeContentQuarantined, processingIsolated, originalPreserved, actualTypeDetected,
+      structurePreserved, extractionDerived, processingMethodsVersioned,
+      normalizationMeaningPreserved, metadataProvenance, classificationPropagated,
+      sensitiveDataGoverned, chunkBoundariesPreserved, contentAwareChunking,
+      aiEnrichmentDerived, curatorAuthorized, separationOfDuties, qualityGatesRequired,
+      publicationHandoffOnly, processingProvenance, reprocessingControlled, retryBounded,
+      poisonContentIsolated, tenantPropertyIsolation, temporaryDataDeleted, providerGoverned,
+      vendorNeutral, technologyIndependent, ingestionActivatesKnowledge, uploadGrantsPublication,
+      duplicateMergesScopes, failedSyncMeansDeletion, extensionDefinesType, extractionIsSourceTruth,
+      normalizationChangesMeaning, classificationBecomesLessRestrictive, chunksCrossBoundaries,
+      oneChunkStrategy, aiChangesAuthority, aiPublishesItself, curatorSelfApprovesCritical,
+      readinessMeansActivation, reprocessingReplacesActive, retriesUnsafeContent,
+      externalProviderDefinesMeaning, directDatabaseAccess, selectsProcessingProvider
+    })) this[key] = Boolean(value);
+    Object.freeze(this);
+  }
+}